# api/src/main/resources/api-definition.yaml
openapi: 3.0.3
info:
  contact:
    name: Sandro Andreasik
  description: Sprint Planning Poker for Scrum teams
  title: EFI Poker API
  version: 1.0.0
servers:
- description: API base path
  url: /api/v1
...
paths:
...
  /projects/{slug}:
    delete:
      operationId: deleteProject
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Project deleted
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Delete project and all associated data (admin)
      tags:
      - projects
    get:
      operationId: getProjectBySlug
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
          description: Project info
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Get project info
      tags:
      - projects
    patch:
      operationId: updateProject
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAdminResponse"
          description: Updated project
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Update project (admin)
      tags:
      - projects
...
  /projects/{slug}/participants:
    get:
      operationId: listProjectParticipants
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ParticipantResponse"
                type: array
          description: List of participants
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: List project participants (admin only)
      tags:
      - participants
    post:
      operationId: joinProject
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinProjectRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantResponse"
          description: Participant info
      summary: Join project (create/get participant by nickname)
      tags:
      - participants
  /projects/{slug}/participants/{participantId}:
    delete:
      operationId: deleteParticipant
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: path
        name: participantId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Participant deleted
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Participant not found
      summary: Delete participant (admin only)
      tags:
      - participants
    get:
      operationId: getParticipant
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: path
        name: participantId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantResponse"
          description: Participant info
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Participant not found
      summary: Get participant by ID (magic link validation)
      tags:
      - participants
    patch:
      operationId: updateParticipant
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: path
        name: participantId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Participant-Id
        required: false
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateParticipantRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantResponse"
          description: Updated participant
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not allowed to update this participant
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Participant not found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Nickname already taken in this project
      summary: Update participant nickname (self-service)
      tags:
      - participants
...
  /rooms/{roomId}:
    get:
      operationId: getRoom
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Participant-Id
        required: false
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetailResponse"
          description: Room details
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Get room with tasks and own estimates
      tags:
      - rooms
    patch:
      operationId: updateRoom
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
          description: Updated room
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
      summary: Update room
      tags:
      - rooms
...
  /rooms/{roomId}/tasks:
    post:
      operationId: createTask
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
          description: Task created
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
      summary: Create a task
      tags:
      - tasks
  /rooms/{roomId}/tasks/import:
    post:
      operationId: importTasks
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportTasksRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/TaskResponse"
                type: array
          description: Tasks imported
      summary: Bulk import tasks
      tags:
      - tasks
...
  /tasks/{taskId}:
    delete:
      operationId: deleteTask
      parameters:
      - in: path
        name: taskId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Task deleted
      summary: Delete a task
      tags:
      - tasks
    patch:
      operationId: updateTask
      parameters:
      - in: path
        name: taskId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
          description: Task updated
      summary: Update a task
      tags:
      - tasks
  /tasks/{taskId}/final-estimate:
    put:
      operationId: setFinalEstimate
      parameters:
      - in: path
        name: taskId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetFinalEstimateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
          description: Final estimate set
      summary: Set or update the final agreed story points for a task
      tags:
      - tasks
...
components:
...
  schemas:
...
    ProjectAdminResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        adminCode: adminCode
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        slug: slug
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        slug:
          type: string
        adminCode:
          type: string
        createdAt:
          format: date-time
          type: string
      type: object
    ProjectResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        slug: slug
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        slug:
          type: string
        createdAt:
          format: date-time
          type: string
      type: object
    ProblemDetail:
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      type: object
    UpdateProjectRequest:
      example:
        name: name
      properties:
        name:
          maxLength: 255
          minLength: 1
          type: string
      type: object
    RoomResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        roundNumber: 0
        description: description
        topic: topic
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        deadline: 2000-01-23T04:56:07.000+00:00
        projectId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        slug: slug
        roomType: ASYNC
        status: OPEN
      properties:
        id:
          format: uuid
          type: string
        slug:
          description: Human-readable room code (e.g. A3X-K7B)
          type: string
        projectId:
          format: uuid
          type: string
        title:
          type: string
        description:
          type: string
        roomType:
          $ref: "#/components/schemas/RoomType"
        deadline:
          format: date-time
          type: string
        status:
          $ref: "#/components/schemas/RoomStatus"
        topic:
          type: string
        roundNumber:
          type: integer
        createdAt:
          format: date-time
          type: string
      type: object
...
    ParticipantResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        nickname: nickname
        invitedRoomIds:
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
        nickname:
          type: string
        invitedRoomIds:
          description: Rooms this participant can access. Empty array means project-wide
            access.
          items:
            format: uuid
            type: string
          type: array
        createdAt:
          format: date-time
          type: string
      type: object
    JoinProjectRequest:
      example:
        nickname: nickname
        roomId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        nickname:
          maxLength: 100
          minLength: 1
          type: string
        roomId:
          description: "When provided, adds access to this room. Multiple room joins\
            \ accumulate access. Project-wide join (no roomId) clears room restrictions."
          format: uuid
          type: string
      required:
      - nickname
      type: object
    UpdateParticipantRequest:
      example:
        nickname: nickname
      properties:
        nickname:
          maxLength: 100
          minLength: 1
          type: string
      required:
      - nickname
      type: object
...
    RoomDetailResponse:
      example:
        roundNumber: 0
        description: description
        topic: topic
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        deadline: 2000-01-23T04:56:07.000+00:00
        slug: slug
        roomType: ASYNC
        tasks:
        - averagePoints: 1.4658129805029452
          allEstimates:
          - participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            storyPoints: storyPoints
            createdAt: 2000-01-23T04:56:07.000+00:00
            participantNickname: participantNickname
            comment: comment
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            storyPoints: storyPoints
            createdAt: 2000-01-23T04:56:07.000+00:00
            participantNickname: participantNickname
            comment: comment
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          questionVotesCount: 2
          description: description
          active: true
          title: title
          finalEstimate: finalEstimate
          sortOrder: 6
          revealed: true
          votedCount: 5
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          totalParticipants: 7
          myEstimate:
            participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            storyPoints: storyPoints
            createdAt: 2000-01-23T04:56:07.000+00:00
            participantNickname: participantNickname
            comment: comment
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          medianPoints: 5.962133916683182
        - averagePoints: 1.4658129805029452
          allEstimates:
          - participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            storyPoints: storyPoints
            createdAt: 2000-01-23T04:56:07.000+00:00
            participantNickname: participantNickname
            comment: comment
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            storyPoints: storyPoints
            createdAt: 2000-01-23T04:56:07.000+00:00
            participantNickname: participantNickname
            comment: comment
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          questionVotesCount: 2
          description: description
          active: true
          title: title
          finalEstimate: finalEstimate
          sortOrder: 6
          revealed: true
          votedCount: 5
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          totalParticipants: 7
          myEstimate:
            participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            storyPoints: storyPoints
            createdAt: 2000-01-23T04:56:07.000+00:00
            participantNickname: participantNickname
            comment: comment
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          medianPoints: 5.962133916683182
        status: OPEN
      properties:
        id:
          format: uuid
          type: string
        slug:
          description: Human-readable room code (e.g. A3X-K7B)
          type: string
        title:
          type: string
        description:
          type: string
        roomType:
          $ref: "#/components/schemas/RoomType"
        deadline:
          format: date-time
          type: string
        status:
          $ref: "#/components/schemas/RoomStatus"
        topic:
          type: string
        roundNumber:
          type: integer
        tasks:
          items:
            $ref: "#/components/schemas/TaskWithEstimateResponse"
          type: array
      type: object
    UpdateRoomRequest:
      example:
        description: description
        topic: topic
        title: title
        deadline: 2000-01-23T04:56:07.000+00:00
      properties:
        title:
          maxLength: 255
          type: string
        description:
          type: string
        deadline:
          format: date-time
          type: string
        topic:
          maxLength: 500
          type: string
      type: object
...
    CreateTaskRequest:
      example:
        sortOrder: 0
        description: description
        title: title
      properties:
        title:
          maxLength: 255
          minLength: 1
          type: string
        description:
          type: string
        sortOrder:
          default: 0
          type: integer
      required:
      - title
      type: object
    TaskResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        sortOrder: 0
        revealed: true
        description: description
        active: true
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        finalEstimate: finalEstimate
        roomId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
        roomId:
          format: uuid
          type: string
        title:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
        finalEstimate:
          type: string
        revealed:
          type: boolean
        active:
          type: boolean
        createdAt:
          format: date-time
          type: string
      type: object
    ImportTasksRequest:
      example:
        titles:
        - titles
        - titles
      properties:
        titles:
          items:
            maxLength: 255
            minLength: 1
            type: string
          minItems: 1
          type: array
      required:
      - titles
      type: object
...
    UpdateTaskRequest:
      example:
        sortOrder: 0
        description: description
        title: title
      properties:
        title:
          maxLength: 255
          type: string
        description:
          type: string
        sortOrder:
          type: integer
      type: object
    SetFinalEstimateRequest:
      example:
        storyPoints: "0"
      properties:
        storyPoints:
          $ref: "#/components/schemas/StoryPoints"
      required:
      - storyPoints
      type: object
...
    EstimateResponse:
      example:
        participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        storyPoints: storyPoints
        createdAt: 2000-01-23T04:56:07.000+00:00
        participantNickname: participantNickname
        comment: comment
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
        participantId:
          format: uuid
          type: string
        participantNickname:
          type: string
        storyPoints:
          type: string
        comment:
          type: string
        createdAt:
          format: date-time
          type: string
      type: object
...
    RoomType:
      enum:
      - ASYNC
      - LIVE
      type: string
    RoomStatus:
      enum:
      - OPEN
      - REVEALED
      - CLOSED
      type: string
    TaskWithEstimateResponse:
      example:
        averagePoints: 1.4658129805029452
        allEstimates:
        - participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          storyPoints: storyPoints
          createdAt: 2000-01-23T04:56:07.000+00:00
          participantNickname: participantNickname
          comment: comment
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          storyPoints: storyPoints
          createdAt: 2000-01-23T04:56:07.000+00:00
          participantNickname: participantNickname
          comment: comment
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        questionVotesCount: 2
        description: description
        active: true
        title: title
        finalEstimate: finalEstimate
        sortOrder: 6
        revealed: true
        votedCount: 5
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        totalParticipants: 7
        myEstimate:
          participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          storyPoints: storyPoints
          createdAt: 2000-01-23T04:56:07.000+00:00
          participantNickname: participantNickname
          comment: comment
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        medianPoints: 5.962133916683182
      properties:
        id:
          format: uuid
          type: string
        title:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
        myEstimate:
          $ref: "#/components/schemas/EstimateResponse"
        allEstimates:
          items:
            $ref: "#/components/schemas/EstimateResponse"
          type: array
        averagePoints:
          format: double
          type: number
        medianPoints:
          format: double
          type: number
        finalEstimate:
          type: string
        votedCount:
          type: integer
        questionVotesCount:
          type: integer
        totalParticipants:
          type: integer
        revealed:
          type: boolean
        active:
          type: boolean
      type: object
...
    StoryPoints:
      enum:
      - "0"
      - "0.5"
      - "1"
      - "2"
      - "3"
      - "5"
      - "8"
      - "13"
      - "21"
      - '?'
      type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http